+++ SignalBox/Buttons.h
@@ -52,7 +52,7 @@ void calibrateButtons();
 /** Read the input button pressed.
  *  Return one of the constants above.
  */
-int readButton();
+uint8_t readButton();
 
 
 /** Wait for button to be released.
@@ -62,7 +62,7 @@ void waitForButtonRelease();
 /** Wait for a button to be pressed.
  *  Return that button after waiting for the release.
  */
-int waitForButton();
+uint8_t waitForButton();
 
 
 #endif
