+++ SignalBox/Report.ino
@@ -13,7 +13,7 @@
 
 /** Is reporting enabled (at a particular level)?
  */
-boolean reportEnabled(int aLevel)
+boolean reportEnabled(uint8_t aLevel)
 {
     return aLevel <= systemData.reportLevel;
 }
